tiny smt opt (#1196)
* tiny smt opt

* restore original test's datasize
