+++ smt/pkg/smt/smt.go
@@ -170,11 +170,7 @@ func (s *SMT) InsertStorage(ethAddr string, storage *map[string]string, chm *map
 		NewRootScalar: &or,
 	}
 	for k := range *storage {
-		keyStoragePosition, err := utils.KeyContractStorage(add, k)
-		if err != nil {
-			return nil, err
-		}
-
+		keyStoragePosition := utils.KeyContractStorage(add, k)
 		smtr, err = s.insert(keyStoragePosition, *(*chm)[k], (*vhm)[k], *smtr.NewRootScalar)
 		if err != nil {
 			return nil, err
@@ -538,14 +534,12 @@ func (s *SMT) insert(k utils.NodeKey, v utils.NodeValue8, newValH [4]uint64, old
 }
 
 func prepareHashValueForSave(in [8]uint64, capacity [4]uint64) utils.NodeValue12 {
-	var sl []uint64
-	sl = append(sl, in[:]...)
-	sl = append(sl, capacity[:]...)
-
 	v := utils.NodeValue12{}
-	for i, val := range sl {
-		b := new(big.Int)
-		v[i] = b.SetUint64(val)
+	for i, val := range in {
+		v[i] = new(big.Int).SetUint64(val)
+	}
+	for i, val := range capacity {
+		v[i+8] = new(big.Int).SetUint64(val)
 	}
 
 	return v
@@ -561,20 +555,12 @@ func (s *SMT) hashSave(in [8]uint64, capacity, h [4]uint64) error {
 }
 
 func (s *SMT) hashcalcAndSave(in [8]uint64, capacity [4]uint64) ([4]uint64, error) {
-	h, err := utils.Hash(in, capacity)
-	if err != nil {
-		return [4]uint64{}, err
-	}
-
+	h := utils.Hash(in, capacity)
 	return h, s.hashSave(in, capacity, h)
 }
 
 func hashCalcAndPrepareForSave(in [8]uint64, capacity [4]uint64) ([4]uint64, utils.NodeValue12, error) {
-	h, err := utils.Hash(in, capacity)
-	if err != nil {
-		return [4]uint64{}, utils.NodeValue12{}, err
-	}
-
+	h := utils.Hash(in, capacity)
 	return h, prepareHashValueForSave(in, capacity), nil
 }
